mod options, locales, statistics
