+++ Locales/pt-br.txt
@@ -1,5 +1,8 @@
 ﻿DEBUGPROFILE	DEBUGGING / PROFILING ONLY
 optionEnableNewTransferManager	ENABLE/DISABLE TRANSFER MANAGER
+GROUP_GENERAL	General Options
+optionPathfindChirper	Chirp about routing/pathfind issues
+optionPathfindChirper_txt	Will regularly chirp about issues when transfers cannot reach their target due to routing issue with your road network. Click the Cim who chirped to find out which building he works in or lives in. Will also display the custom building name if present (internal model name otherwise). THE BIRD ONLY NOTIFIES - YOU HAVE TO FIX THE PROBLEMS!
 GROUP_SERVICE_OPTIONS	Opções de Serviço
 optionPreferLocalService	Preferir serviços no distrito local
 optionPreferLocalService_txt	Melhora ainda mais os serviços locais, conectando requisições de prioridade baixa somente com serviços dentro do mesmo distrito (ou fora de qualquer distrito). Se uma requisição se tornar de alta prioridade, ela será atendida por qualquer serviço, independente do distrito que está. Afeta: lixo, polícia, saúde, manutenção, email e taxi.
