+++ MoreEffectiveTransfer.csproj
@@ -84,6 +84,7 @@
     <Compile Include="CustomManager\CustomTransferManager.cs" />
     <Compile Include="Patch\CarAIPathfindFailurePatch.cs" />
     <Compile Include="Patch\TransferManagerMatchOfferPatch.cs" />
+    <Compile Include="UI\Settings.cs" />
     <Compile Include="Util\PathFindFailure.cs" />
     <Compile Include="Util\Profiling.cs" />
     <Compile Include="Util\FastList.cs" />
@@ -105,7 +106,6 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="CustomAI\" />
-    <Folder Include="UI\" />
   </ItemGroup>
   <ItemGroup>
     <None Include="CHANGELOG" />
